More Algorithms, DGML Output of Binary Trees

+++ InterviewC/DynamicProgramming.cpp
@@ -1,7 +1,9 @@
+#include <cstdint>
+
 bool can_win(int* a, size_t n, int cur, int sum) {
 	if (cur >= sum) return false;
 
-	for (int i = 0; i < n; i++) {
+	for (size_t i = 0; i < n; i++) {
 		if (!can_win(a, n, cur + a[i], sum)) {
 			//printf("%d\n", a[i]);
 			return true;
